style(fmt): batch 17 - IDE0004/0040/0055/0090/0200 (cast/modifier/format)
Apply 'dotnet format style' for these style rules:
- IDE0004 remove unnecessary casts
- IDE0040 add accessibility modifier where missing
- IDE0055 fix formatting
- IDE0090 use target-typed 'new()'
- IDE0200 remove unnecessary lambda expression

Verified: build with analyzers 0/0.

+++ Tests/Opc.Ua.PubSub.Tests/Encoding/PubSubJsonEncoderAdditionalTests.cs
@@ -637,7 +637,7 @@ namespace Opc.Ua.PubSub.Tests.Encoding
                     BuiltInType = (byte)BuiltInType.UInt32,
                     ValueRank = ValueRanks.Scalar
                 },
-                Value = new DataValue(new Variant((uint)4000000000))
+                Value = new DataValue(new Variant(4000000000))
             };
 
             var message = new PubSubEncoding.JsonDataSetMessage(
